--- src/components/App.css
@@ -1,162 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-*::-webkit-scrollbar {
-  display: none;
-}
-
-body {
-  margin: 0;
-  font-family: Pretendard;
-  background-color: #ffffff;
-}
-
-.App {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #f0f6ff;
-}
-
-.App-nav {
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 0;
-  width: 100%;
-  background-color: #f0f6ff;
-}
-
-.App-nav-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 6.3rem;
-  padding: 0 3.2rem;
-
-  @media (min-width: 768px) {
-    height: 9.4rem;
-    max-width: 86.3rem;
-  }
-
-  @media (min-width: 1200px) {
-    height: 9.4rem;
-    max-width: 192rem;
-    padding: 0 20rem;
-  }
-}
-
-.logo {
-  width: 8.3125rem;
-  height: 1.5rem;
-
-  @media (min-width: 768px) {
-    height: 2.4rem;
-  }
-}
-
-.profile {
-  display: flex;
-  color: #373740;
-  font-size: 14px;
-  font-weight: 400;
-  text-decoration: none;
-  align-items: center;
-  gap: 6px;
-}
-
-.profileImg {
-  width: 1.75rem;
-  height: 1.75rem;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.loginBtn {
-  display: flex;
-  width: 8rem;
-  padding: 16px 20px;
-  justify-content: center;
-  align-items: center;
-  background-image: linear-gradient(135deg, var(--primary) 0%, #6ae3fe 100%);
-  cursor: pointer;
-  border-radius: 0.8rem;
-  color: #f5f5f5;
-  font-size: 1.125rem;
-  font-weight: 600;
-}
-
-header {
-  display: flex;
-  width: 100%;
-  padding: 1.25rem 0 3.75rem;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-}
-
-.user {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: #000000;
-  text-align: center;
-  font-size: 40px;
-  font-weight: 600;
-  gap: 20px;
-}
-
-.myInfo {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 0.875rem;
-  font-weight: 400;
-  line-height: 1.5rem;
-  gap: 12px;
-}
-
-.myImgBg {
-  display: flex;
-  height: 60px;
-  padding: 16px 17px 16px 15px;
-  border-radius: 47px;
-  background: #8f00ff;
-  justify-content: center;
-  align-items: center;
-}
-
-.myImg {
-  width: 28px;
-  height: 28px;
-}
-
-.App-container {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  gap: 40px;
-  background-color: #ffffff;
-}
-
-.search {
-  margin: 40px auto 0;
-  width: 70%;
-  max-width: 1050px;
-  height: 3rem;
-  padding: 0.9375rem 1rem;
-  border-radius: 10px;
-  background-color: #f5f5f5;
-}
-
-footer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  height: 13rem;
-  background-color: #111322;
-}
