+++ src/css/CardList.css
@@ -33,6 +33,7 @@
 
 .card:hover img {
   transform: scale(1.3);
+  z-index: 1;
 }
 
 .card a {
@@ -47,6 +48,11 @@
   flex-shrink: 0;
   object-fit: contain;
   transition: transform 0.5s;
+}
+
+.cardImgContainer {
+  height: 12.5rem;
+  overflow: hidden;
 }
 
 a {
