+++ src/css/Footer.css
@@ -1,3 +1,11 @@
+footer {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  height: 13rem;
+  background-color: #111322;
+}
+
 .footer-box {
   display: grid;
   justify-content: space-between;
