+++ src/css/Profile.css
@@ -1,0 +1,30 @@
+.profile {
+  display: flex;
+  color: #373740;
+  font-size: 14px;
+  font-weight: 400;
+  text-decoration: none;
+  align-items: center;
+  gap: 6px;
+}
+
+.profileImg {
+  width: 1.75rem;
+  height: 1.75rem;
+  border-radius: 50%;
+  flex-shrink: 0;
+}
+
+.loginBtn {
+  display: flex;
+  width: 8rem;
+  padding: 16px 20px;
+  justify-content: center;
+  align-items: center;
+  background-image: linear-gradient(135deg, var(--primary) 0%, #6ae3fe 100%);
+  cursor: pointer;
+  border-radius: 0.8rem;
+  color: #f5f5f5;
+  font-size: 1.125rem;
+  font-weight: 600;
+}
