+++ src/css/Search.css
@@ -1,0 +1,9 @@
+.search {
+  margin: 40px auto 0;
+  width: 70%;
+  max-width: 1050px;
+  height: 3rem;
+  padding: 0.9375rem 1rem;
+  border-radius: 10px;
+  background-color: #f5f5f5;
+}
